Consolidating face detection view controller into one file.
